Take 2 of a better way to share logic between Local and Proxy servers
