+++ packages/core/src/initVal.ts
@@ -25,6 +25,7 @@ export type ValConstructor = {
 
 export type ValConfig = {
   remote?: string;
+  root?: string;
   gitCommit?: string;
   gitBranch?: string;
   valConfigPath?: string;
