+++ packages/server/src/LocalValServer.ts
@@ -1,28 +1,17 @@
 import { Service } from "./Service";
 import { result } from "@valbuild/core/fp";
-import { JSONOps, JSONValue, applyPatch } from "@valbuild/core/patch";
 import { Patch } from "./patch/validation";
 import {
-  ApiCommitResponse,
   ApiGetPatchResponse,
   ApiPostPatchResponse,
-  ApiPostValidationErrorResponse,
-  ApiTreeResponse,
   ModuleId,
   PatchId,
   ApiDeletePatchResponse,
-  deserializeSchema,
-  SourcePath,
-  ApiPostValidationResponse,
-  ValidationError,
-  FileMetadata,
-  ImageMetadata,
 } from "@valbuild/core";
 import {
   VAL_ENABLE_COOKIE_NAME,
   VAL_SESSION_COOKIE,
   VAL_STATE_COOKIE,
-  ValCookies,
   ValServerError,
   ValServerJsonResult,
   ValServerRedirectResult,
@@ -31,19 +20,9 @@ import {
 } from "@valbuild/shared/internal";
 import path from "path";
 import { z } from "zod";
-import {
-  ValServer,
-  getRedirectUrl,
-  ENABLE_COOKIE_VALUE,
-  ValServerCallbacks,
-} from "./ValServer";
-import { SerializedModuleContent } from "./SerializedModuleContent";
-import { getValidationErrorFileRef } from "./getValidationErrorFileRef";
-import { extractFileMetadata, extractImageMetadata } from "./extractMetadata";
-import { validateMetadata } from "./validateMetadata";
-import { getValidationErrorMetadata } from "./getValidationErrorMetadata";
-import { IValFSHost } from "./ValFSHost";
+import { ValServer, ValServerCallbacks } from "./ValServer";
 import fs from "fs";
+import { SerializedModuleContent } from "./SerializedModuleContent";
 
 export type LocalValServerOptions = {
   service: Service;
@@ -56,29 +35,26 @@ export type LocalValServerOptions = {
   };
 };
 
-const ops = new JSONOps();
-
 export interface ValServerBufferHost {
   readBuffer: (fileName: string) => Promise<Buffer | undefined>;
 }
 
-export class LocalValServer implements ValServer {
+export class LocalValServer extends ValServer {
   private static readonly PATCHES_DIR = "patches";
   private readonly patchesRootPath: string;
-  private readonly host: IValFSHost;
   constructor(
     readonly options: LocalValServerOptions,
-    readonly callbacks: ValServerCallbacks,
-    readonly bufferHost: ValServerBufferHost = {
-      readBuffer: (fileName: string) => {
-        return fs.promises.readFile(fileName);
-      },
-    }
+    readonly callbacks: ValServerCallbacks
   ) {
+    super(
+      options.service.sourceFileHandler.projectRoot,
+      options.service.sourceFileHandler.host,
+      options,
+      callbacks
+    );
     this.patchesRootPath =
       options.cacheDir ||
       path.join(options.service.sourceFileHandler.projectRoot, ".val");
-    this.host = options.service.sourceFileHandler.host;
   }
 
   async session(): Promise<ValServerJsonResult<ValSession>> {
@@ -90,298 +66,109 @@ export class LocalValServer implements ValServer {
       },
     };
   }
-
-  async getTree(
-    treePath: string,
-    // TODO: use the params: patch, schema, source
-    query: { patch?: string; schema?: string; source?: string },
-    cookies: ValCookies<VAL_SESSION_COOKIE>
-  ): Promise<ValServerJsonResult<ApiTreeResponse>> {
-    const rootDir = process.cwd();
-    const moduleIds: ModuleId[] = this.host
-      .readDirectory(
-        rootDir,
-        ["ts", "js"],
-        ["node_modules", ".*"],
-        ["**/*.val.ts", "**/*.val.js"]
-      )
-      .filter((file) => {
-        if (treePath) {
-          return file.replace(rootDir, "").startsWith(treePath);
-        }
-        return true;
-      })
-      .map(
-        (file) =>
-          file
-            .replace(rootDir, "")
-            .replace(".val.js", "")
-            .replace(".val.ts", "")
-            .split(path.sep)
-            .join("/") as ModuleId
+  async deletePatches(query: {
+    id?: string[];
+  }): Promise<ValServerJsonResult<ApiDeletePatchResponse>> {
+    const deletedPatches: ApiDeletePatchResponse = [];
+    for (const patchId of query.id ?? []) {
+      deletedPatches.push(patchId as PatchId);
+      this.host.rmFile(
+        path.join(this.patchesRootPath, LocalValServer.PATCHES_DIR, patchId)
       );
-    const applyPatches = query.patch === "true";
-    let {
-      patchIdsByModuleId,
-      patchesById,
-    }: {
-      patchIdsByModuleId: Record<ModuleId, PatchId[]>;
-      patchesById: Record<PatchId, Patch>;
-    } = {
-      patchIdsByModuleId: {},
-      patchesById: {},
-    };
-    if (applyPatches) {
-      const res = await this.readPatches(cookies);
-      if (result.isErr(res)) {
-        return res.error;
-      }
-      patchIdsByModuleId = res.value.patchIdsByModuleId;
-      patchesById = res.value.patchesById;
     }
-
-    const possiblyPatchesContent = await Promise.all(
-      moduleIds.map(async (moduleId) => {
-        return this.applyAllPatchesThenValidate(
-          moduleId,
-          patchIdsByModuleId,
-          patchesById,
-          applyPatches,
-          cookies
-        );
-      })
-    );
-
-    const modules = Object.fromEntries(
-      possiblyPatchesContent.map((serializedModuleContent) => {
-        const module: ApiTreeResponse["modules"][keyof ApiTreeResponse["modules"]] =
-          {
-            schema: serializedModuleContent.schema,
-            source: serializedModuleContent.source,
-            errors: serializedModuleContent.errors,
-          };
-        return [serializedModuleContent.path, module];
-      })
-    );
-    const apiTreeResponse: ApiTreeResponse = {
-      modules,
-      git: this.options.git,
-    };
     return {
       status: 200,
-      json: apiTreeResponse,
+      json: deletedPatches,
     };
   }
 
-  private async applyAllPatchesThenValidate(
-    moduleId: ModuleId,
-    patchIdsByModuleId: Record<ModuleId, PatchId[]>,
-    patchesById: Record<PatchId, Patch>,
-    applyPatches: boolean,
-    cookies: ValCookies<VAL_SESSION_COOKIE>
-  ): Promise<SerializedModuleContent> {
-    const serializedModuleContent = await this.options.service.get(moduleId);
-    const schema = serializedModuleContent.schema;
-    const maybeSource = serializedModuleContent.source;
-    if (!applyPatches) {
-      return serializedModuleContent;
-    }
-    if (
-      serializedModuleContent.errors &&
-      (serializedModuleContent.errors.fatal ||
-        serializedModuleContent.errors.invalidModuleId)
-    ) {
-      return serializedModuleContent;
-    }
-    if (!maybeSource || !schema) {
-      return serializedModuleContent;
-    }
-    let source = maybeSource as JSONValue;
-
-    for (const patchId of patchIdsByModuleId[moduleId] ?? []) {
-      const patch = patchesById[patchId];
-      if (!patch) {
-        continue;
-      }
-      const patchRes = applyPatch(source, ops, patch);
-      if (result.isOk(patchRes)) {
-        source = patchRes.value;
-      } else {
-        console.error(
-          "Val: got an unexpected error while applying patch. Is there a mismatch in Val versions? Perhaps Val is misconfigured?",
-          {
-            patchId,
-            moduleId,
-            patch,
-            error: patchRes.error,
-          }
-        );
-        return {
-          path: moduleId as string as SourcePath,
-          schema,
-          source,
-          errors: {
-            fatal: [
-              {
-                message: "Unexpected error applying patch",
-                type: "invalid-patch",
-              },
-            ],
-          },
-        };
-      }
-    }
-
-    const validationErrors = deserializeSchema(schema).validate(
-      moduleId as string as SourcePath,
-      source
-    );
-    if (validationErrors) {
-      const revalidated = await this.revalidateImageAndFileValidation(
-        validationErrors,
-        cookies
-      );
+  async postPatches(
+    body: unknown
+  ): Promise<ValServerJsonResult<ApiPostPatchResponse>> {
+    const patches = z.record(Patch).safeParse(body);
+    if (!patches.success) {
       return {
-        path: moduleId as string as SourcePath,
-        schema,
-        source,
-        errors: revalidated && {
-          validation: revalidated,
+        status: 404,
+        json: {
+          message: `Invalid patch: ${patches.error.message}`,
+          details: patches.error.issues,
         },
       };
     }
-
-    return {
-      path: moduleId as string as SourcePath,
-      schema,
-      source,
-      errors: false,
-    };
-  }
-
-  protected async getMetadata(
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    filePath: string,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    cookies: ValCookies<VAL_SESSION_COOKIE>
-  ): Promise<
-    | { status: 200; json: ImageMetadata | FileMetadata | undefined }
-    | ValServerError
-  > {
+    const parsedPatches: Record<ModuleId, Patch> = {};
+    const fileId = Date.now().toString() as PatchId;
+    const res: ApiPostPatchResponse = {};
+    for (const moduleIdStr in patches.data) {
+      const moduleId = moduleIdStr as ModuleId; // TODO: validate that this is a valid module id
+      res[moduleId] = {
+        patch_id: fileId.toString() as PatchId,
+      };
+      parsedPatches[moduleId] = patches.data[moduleId];
+    }
+    this.host.writeFile(
+      this.getPatchFilePath(fileId),
+      JSON.stringify(patches.data),
+      "utf8"
+    );
     return {
       status: 200,
-      json: undefined,
+      json: res,
     };
   }
 
-  // TODO: name this better: we need to check for image and file validation errors
-  // since they cannot be handled directly inside the validation function.
-  // The reason is that validate will be called inside QuickJS (in the future, hopefully), which does not have access to the filesystem.
-  // If you are reading this, and we still are not using QuickJS to validate, this assumption might be wrong.
-  private async revalidateImageAndFileValidation(
-    validationErrors: Record<SourcePath, ValidationError[]>,
-    cookies: ValCookies<VAL_SESSION_COOKIE>
-  ): Promise<false | Record<SourcePath, ValidationError[]>> {
-    const revalidatedValidationErrors: Record<SourcePath, ValidationError[]> =
-      {};
-    for (const pathStr in validationErrors) {
-      const errorSourcePath = pathStr as SourcePath;
-      const errors = validationErrors[errorSourcePath];
-      revalidatedValidationErrors[errorSourcePath] = [];
-      for (const error of errors) {
-        if (
-          error.fixes?.every(
-            (fix) =>
-              fix === "file:check-metadata" || fix === "image:replace-metadata" // TODO: rename fix to: image:check-metadata
-          )
-        ) {
-          const fileRef = getValidationErrorFileRef(error);
-          if (fileRef) {
-            const filePath = path.join(
-              this.options.service.sourceFileHandler.projectRoot,
-              fileRef
+  async getPatches(query: {
+    id?: string[];
+  }): Promise<ValServerJsonResult<ApiGetPatchResponse>> {
+    const readRes = await this.readPatches();
+    if (result.isErr(readRes)) {
+      return readRes.error;
+    }
+    const res: ApiGetPatchResponse = {};
+    const { patchIdsByModuleId, patchesById } = readRes.value;
+    for (const moduleIdStr in patchIdsByModuleId) {
+      const moduleId = moduleIdStr as ModuleId;
+      if (
+        (query.id && query.id.includes(moduleId)) ||
+        !query.id ||
+        query.id.length === 0
+      ) {
+        res[moduleId] = patchIdsByModuleId[moduleId].map((patchId) => {
+          let createdAt = new Date(0);
+          try {
+            createdAt = new Date(parseInt(patchId, 10));
+          } catch (e) {
+            console.error(
+              "Val: unexpected error parsing patch ids. Is cache corrupt?",
+              {
+                patchId,
+                file: this.getPatchFilePath(patchId),
+                dir: this.getPatchesCacheDir(),
+                error: e,
+              }
             );
-            let expectedMetadata: FileMetadata | ImageMetadata | undefined;
-
-            // if file is not in FS (e.g. remote) we try to get the metadata directly from server
-            const directMetadata = await this.getMetadata(filePath, cookies);
-            if (directMetadata.status === 200) {
-              expectedMetadata = directMetadata.json;
-            } else {
-              throw Error(directMetadata.json?.message);
-            }
-
-            // if this is a new file or we have an actual FS, we read the file and get the metadata
-            if (!expectedMetadata) {
-              let fileBuffer: Buffer | undefined = undefined;
-              try {
-                fileBuffer = await this.bufferHost.readBuffer(filePath);
-              } catch (err) {
-                //
-              }
-              if (!fileBuffer) {
-                revalidatedValidationErrors[errorSourcePath].push({
-                  message: `Could not read file: ${filePath}`,
-                });
-                continue;
-              }
-
-              if (error.fixes.some((fix) => fix === "image:replace-metadata")) {
-                expectedMetadata = await extractImageMetadata(
-                  filePath,
-                  fileBuffer
-                );
-              } else {
-                expectedMetadata = await extractFileMetadata(
-                  filePath,
-                  fileBuffer
-                );
-              }
-            }
-            if (!expectedMetadata) {
-              revalidatedValidationErrors[errorSourcePath].push({
-                message: `Could not read file metadata. Is the reference to the file: ${fileRef} correct?`,
-              });
-            } else {
-              const actualMetadata = getValidationErrorMetadata(error);
-              const revalidatedError = validateMetadata(
-                actualMetadata,
-                expectedMetadata
-              );
-              if (!revalidatedError) {
-                // no errors anymore:
-                continue;
-              }
-              const errorMsgs = (revalidatedError.globalErrors || [])
-                .concat(Object.values(revalidatedError.erroneousMetadata || {}))
-                .concat(
-                  Object.values(revalidatedError.missingMetadata || []).map(
-                    (missingKey) => `Required key: ${missingKey} is not defined`
-                  )
-                );
-              revalidatedValidationErrors[errorSourcePath].push(
-                ...errorMsgs.map((message) => ({ message }))
-              );
-            }
-          } else {
-            revalidatedValidationErrors[errorSourcePath].push(error);
+            throw Error(
+              "Unexpected error parsing patch ids. Is cache corrupt?"
+            );
           }
-        } else {
-          revalidatedValidationErrors[errorSourcePath].push(error);
-        }
+          return {
+            patch_id: patchId,
+            patch: patchesById[patchId],
+            created_at: createdAt.toISOString(),
+          };
+        });
       }
     }
-    const hasErrors = Object.values(revalidatedValidationErrors).some(
-      (errors) => errors.length > 0
-    );
-    if (hasErrors) {
-      return revalidatedValidationErrors;
-    }
-    return hasErrors;
+    return {
+      status: 200,
+      json: res,
+    };
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  protected async readPatches(cookies: ValCookies<VAL_SESSION_COOKIE>): Promise<
+  protected async readBuffer(filePath: string): Promise<Buffer> {
+    return fs.promises.readFile(filePath);
+  }
+
+  protected async readPatches(): Promise<
     result.Result<
       {
         patches: [PatchId, ModuleId, Patch][];
@@ -476,98 +263,6 @@ export class LocalValServer implements ValServer {
     });
   }
 
-  async enable(query: {
-    redirect_to?: string;
-  }): Promise<ValServerRedirectResult<VAL_ENABLE_COOKIE_NAME>> {
-    const redirectToRes = getRedirectUrl(
-      query,
-      this.options.valEnableRedirectUrl
-    );
-    if (typeof redirectToRes !== "string") {
-      return redirectToRes;
-    }
-    await this.callbacks.onEnable(true);
-    return {
-      cookies: {
-        [VAL_ENABLE_COOKIE_NAME]: ENABLE_COOKIE_VALUE,
-      },
-      status: 302,
-      redirectTo: redirectToRes,
-    };
-  }
-
-  async disable(query: {
-    redirect_to?: string;
-  }): Promise<ValServerRedirectResult<VAL_ENABLE_COOKIE_NAME>> {
-    const redirectToRes = getRedirectUrl(
-      query,
-      this.options.valDisableRedirectUrl
-    );
-    if (typeof redirectToRes !== "string") {
-      return redirectToRes;
-    }
-    await this.callbacks.onDisable(true);
-    return {
-      cookies: {
-        [VAL_ENABLE_COOKIE_NAME]: {
-          value: "false",
-        },
-      },
-      status: 302,
-      redirectTo: redirectToRes,
-    };
-  }
-
-  async deletePatches(query: {
-    id?: string[];
-  }): Promise<ValServerJsonResult<ApiDeletePatchResponse>> {
-    const deletedPatches: ApiDeletePatchResponse = [];
-    for (const patchId of query.id ?? []) {
-      deletedPatches.push(patchId as PatchId);
-      this.host.rmFile(
-        path.join(this.patchesRootPath, LocalValServer.PATCHES_DIR, patchId)
-      );
-    }
-    return {
-      status: 200,
-      json: deletedPatches,
-    };
-  }
-
-  async postPatches(
-    body: unknown
-  ): Promise<ValServerJsonResult<ApiPostPatchResponse>> {
-    const patches = z.record(Patch).safeParse(body);
-    if (!patches.success) {
-      return {
-        status: 404,
-        json: {
-          message: `Invalid patch: ${patches.error.message}`,
-          details: patches.error.issues,
-        },
-      };
-    }
-    const parsedPatches: Record<ModuleId, Patch> = {};
-    const fileId = Date.now().toString() as PatchId;
-    const res: ApiPostPatchResponse = {};
-    for (const moduleIdStr in patches.data) {
-      const moduleId = moduleIdStr as ModuleId; // TODO: validate that this is a valid module id
-      res[moduleId] = {
-        patch_id: fileId.toString() as PatchId,
-      };
-      parsedPatches[moduleId] = patches.data[moduleId];
-    }
-    this.host.writeFile(
-      this.getPatchFilePath(fileId),
-      JSON.stringify(patches.data),
-      "utf8"
-    );
-    return {
-      status: 200,
-      json: res,
-    };
-  }
-
   private getPatchesCacheDir() {
     return path.join(this.patchesRootPath, LocalValServer.PATCHES_DIR);
   }
@@ -589,118 +284,30 @@ export class LocalValServer implements ValServer {
     };
   }
 
-  private async validateThenMaybeCommit(
-    rawBody: unknown,
-    commit: boolean,
-    cookies: ValCookies<VAL_SESSION_COOKIE>
-  ): Promise<
-    Promise<
-      ValServerJsonResult<
-        ApiPostValidationResponse | ApiPostValidationErrorResponse
-      >
-    >
+  protected async ensureRemoteFSInitialized(): Promise<
+    result.Result<undefined, ValServerError>
   > {
-    const filterPatchesByModuleIdRes = z
-      .object({
-        patches: z.record(z.array(z.string())).optional(),
-      })
-      .safeParse(rawBody);
-    if (!filterPatchesByModuleIdRes.success) {
-      return {
-        status: 404,
-        json: {
-          message: "Could not parse body",
-          details: filterPatchesByModuleIdRes.error,
-        },
-      };
-    }
-    let {
-      patchIdsByModuleId,
-      patchesById,
-      patches,
-    }: {
-      patchIdsByModuleId: Record<ModuleId, PatchId[]>;
-      patchesById: Record<PatchId, Patch>;
-      patches: [PatchId, ModuleId, Patch][];
-    } = {
-      patchIdsByModuleId: {},
-      patchesById: {},
-      patches: [],
-    };
-    const res = await this.readPatches(cookies);
-    if (result.isErr(res)) {
-      return res.error;
-    }
-    patchIdsByModuleId = res.value.patchIdsByModuleId;
-    patchesById = res.value.patchesById;
-    patches = res.value.patches;
-
-    const validationErrorsByModuleId: ApiPostValidationErrorResponse["validationErrors"] =
-      {};
-    for (const moduleIdStr in patchIdsByModuleId) {
-      const moduleId = moduleIdStr as ModuleId;
-      const serializedModuleContent = await this.applyAllPatchesThenValidate(
-        moduleId,
-        (filterPatchesByModuleIdRes.data.patches as Record<
-          ModuleId,
-          PatchId[]
-        >) || // TODO: refine to ModuleId and PatchId when parsing
-          patchIdsByModuleId,
-        patchesById,
-        true,
-        cookies
-      );
-      if (serializedModuleContent.errors) {
-        validationErrorsByModuleId[moduleId] = serializedModuleContent;
-      }
-    }
-    if (Object.keys(validationErrorsByModuleId).length > 0) {
-      const modules: Record<
-        ModuleId,
-        {
-          patches: {
-            applied: PatchId[];
-            failed?: PatchId[];
-          };
-        }
-      > = {};
-      for (const [patchId, moduleId] of patches) {
-        if (!modules[moduleId]) {
-          modules[moduleId] = {
-            patches: {
-              applied: [],
-            },
-          };
-        }
-        if (validationErrorsByModuleId[moduleId]) {
-          if (!modules[moduleId].patches.failed) {
-            modules[moduleId].patches.failed = [];
-          }
-          modules[moduleId].patches.failed?.push(patchId);
-        } else {
-          modules[moduleId].patches.applied.push(patchId);
-        }
-      }
-      return {
-        status: 200,
-        json: {
-          modules,
-          validationErrors: validationErrorsByModuleId,
-        },
-      };
-    }
-
-    const modules = await this.getModulesWithAppliedPatches(commit, patches);
-    return {
-      status: 200,
-      json: {
-        modules,
-        validationErrors: false,
-      },
-    };
+    // No RemoteFS so nothing to ensure
+    return result.ok(undefined);
   }
 
-  protected async getModulesWithAppliedPatches(
+  protected getModule(moduleId: ModuleId): Promise<SerializedModuleContent> {
+    return this.options.service.get(moduleId);
+  }
+
+  protected getPatchedModules(
+    patches: [PatchId, ModuleId, Patch][]
+  ): Promise<Record<ModuleId, { patches: { applied: PatchId[] } }>> {
+    return this.commitOrGetModulesWithAppliedPatches(false, patches);
+  }
+
+  protected execCommit(
+    patches: [PatchId, ModuleId, Patch][]
+  ): Promise<Record<ModuleId, { patches: { applied: PatchId[] } }>> {
+    return this.commitOrGetModulesWithAppliedPatches(true, patches);
+  }
+
+  private async commitOrGetModulesWithAppliedPatches(
     commit: boolean,
     patches: [PatchId, ModuleId, Patch][]
   ) {
@@ -734,46 +341,7 @@ export class LocalValServer implements ValServer {
     return modules;
   }
 
-  async postValidate(
-    rawBody: unknown,
-    cookies: ValCookies<VAL_SESSION_COOKIE>
-  ): Promise<
-    ValServerJsonResult<
-      ApiPostValidationResponse | ApiPostValidationErrorResponse
-    >
-  > {
-    return this.validateThenMaybeCommit(rawBody, false, cookies);
-  }
-
-  async postCommit(
-    rawBody: unknown,
-    cookies: ValCookies<VAL_SESSION_COOKIE>
-  ): Promise<
-    ValServerJsonResult<ApiCommitResponse, ApiPostValidationErrorResponse>
-  > {
-    const res = await this.validateThenMaybeCommit(rawBody, true, cookies);
-    if (res.status === 200) {
-      if (res.json.validationErrors) {
-        return {
-          status: 400,
-          json: {
-            ...res.json,
-          },
-        } as { status: 400; json: ApiPostValidationErrorResponse };
-      }
-      return {
-        status: 200,
-        json: {
-          ...res.json,
-          git: this.options.git,
-        },
-      };
-    }
-    return res as ValServerJsonResult<
-      ApiCommitResponse,
-      ApiPostValidationErrorResponse
-    >;
-  }
+  /* Bad requests on Local Server: */
 
   async authorize(): Promise<ValServerRedirectResult<VAL_STATE_COOKIE>> {
     return this.badRequest();
@@ -797,56 +365,5 @@ export class LocalValServer implements ValServer {
     ValServerResult<never, ReadableStream<Uint8Array>>
   > {
     return this.badRequest();
-  }
-
-  async getPatches(
-    query: {
-      id?: string[];
-    },
-    cookies: ValCookies<VAL_SESSION_COOKIE>
-  ): Promise<ValServerJsonResult<ApiGetPatchResponse>> {
-    const readRes = await this.readPatches(cookies);
-    if (result.isErr(readRes)) {
-      return readRes.error;
-    }
-    const res: ApiGetPatchResponse = {};
-    const { patchIdsByModuleId, patchesById } = readRes.value;
-    for (const moduleIdStr in patchIdsByModuleId) {
-      const moduleId = moduleIdStr as ModuleId;
-      if (
-        (query.id && query.id.includes(moduleId)) ||
-        !query.id ||
-        query.id.length === 0
-      ) {
-        res[moduleId] = patchIdsByModuleId[moduleId].map((patchId) => {
-          let createdAt = new Date(0);
-          try {
-            createdAt = new Date(parseInt(patchId, 10));
-          } catch (e) {
-            console.error(
-              "Val: unexpected error parsing patch ids. Is cache corrupt?",
-              {
-                patchId,
-                file: this.getPatchFilePath(patchId),
-                dir: this.getPatchesCacheDir(),
-                error: e,
-              }
-            );
-            throw Error(
-              "Unexpected error parsing patch ids. Is cache corrupt?"
-            );
-          }
-          return {
-            patch_id: patchId,
-            patch: patchesById[patchId],
-            created_at: createdAt.toISOString(),
-          };
-        });
-      }
-    }
-    return {
-      status: 200,
-      json: res,
-    };
   }
 }
